Make sure failure to remove temporary files does not cause error exit codes
Summary:
For some reason on Windows, the files used for the streaming stdout / stderr remain locked when PHP exits.  In this case, the automatic unlink fails, and PHP returns an error code.

Because we don't have much control of Windows and how it does it's file locking nonsense, gracefully ignore failure to remove TempFiles if `$temp->setIgnoreRemovalFailure(true)` is called.

Also, have a free "latest XHPAST built for Windows".

Test Plan:
Replicated the issue and then set:

```
$this->windowsStdoutTempFile->setIgnoreRemovalFailure(true);
$this->windowsStderrTempFile->setIgnoreRemovalFailure(true);
```

and saw the failure go away.  Set those values to false and saw the exception appear again.  Set them back to true before running `arc diff` so that we actually fix the issue.

Reviewers: #blessed_reviewers!, epriestley

Subscribers: nickhutchinson, eadler, epriestley, Korvin

Differential Revision: https://secure.phabricator.com/D13685