TempFile: don't create a temporary directory to hold a temporary file in most cases
On Windows a directory may fail to be deleted if it contains a recently
deleted file. Update TempFile to prefer to create a file in the system
temporary directory instead of creating an interstitial directory to
hold the temporary file.

Also fix a dodgy unit test that relied on being able to write to
`dirname(dirname($my_temp_file))`